--- 05_react_app/src/App.css
@@ -1,3 +1,0 @@
-.firstComp {
-  color: red;
-}
