--- 05_react_app/src/FirstComponent.js
@@ -1,6 +1,0 @@
-const FirstComponent  = (props) => {
-    console.log(props);
-    return <div className="firstComp">{props.textToShow}</div>
-}
-
-export default FirstComponent;
