+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "05_react_app",
+  "name": "app_01",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
